+++ rts/motoko-rts/src/memory.rs
@@ -27,17 +27,17 @@ use motoko_rts_macros::ic_mem_fn;
 /// This function does not take any `Memory` arguments can be used by the generated code.
 pub trait Memory {
     // General allocator working for all GC variants.
-    unsafe fn alloc_words(&mut self, n: Words<u32>) -> Value;
+    unsafe fn alloc_words(&mut self, n: Words<usize>) -> Value;
 
     // Grow the allocated memory size to at least the address of `ptr`.
-    unsafe fn grow_memory(&mut self, ptr: u64);
+    unsafe fn grow_memory(&mut self, ptr: usize);
 }
 
 /// Allocate a new blob.
 /// Note: After initialization, the post allocation barrier needs to be applied to all mutator objects.
 /// For RTS-internal blobs that can be collected by the next GC run, the post allocation barrier can be omitted.
 #[ic_mem_fn]
-pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<u32>) -> Value {
+pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<usize>) -> Value {
     let ptr = mem.alloc_words(size_of::<Blob>() + size.to_words());
     // NB. Cannot use `as_blob` here as we didn't write the header yet
     let blob = ptr.get_ptr() as *mut Blob;
@@ -51,7 +51,7 @@ pub unsafe fn alloc_blob<M: Memory>(mem: &mut M, size: Bytes<u32>) -> Value {
 /// Allocate a new array.
 /// Note: After initialization, the post allocation barrier needs to be applied to all mutator objects.
 #[ic_mem_fn]
-pub unsafe fn alloc_array<M: Memory>(mem: &mut M, len: u32) -> Value {
+pub unsafe fn alloc_array<M: Memory>(mem: &mut M, len: usize) -> Value {
     // Array payload should not be larger than half of the memory
     if len > (WASM_HEAP_SIZE / 2).0 {
         rts_trap_with("Array allocation too large");
